requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
pkg-resources==0.0.0
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
Werkzeug==1.0.0
WTForms==2.2.1
